requirements.txt
panel
matplotlib
pandas
shap
xgboost
--extra-index-url=https://pypi.nvidia.com 
cudf-cu12==24.6.* 
dask-cudf-cu12==24.6.* 
cuml-cu12==24.6.*
cugraph-cu12==24.6.* 
cuspatial-cu12==24.6.* 
cuproj-cu12==24.6.*
cuxfilter-cu12==24.6.* 
cucim-cu12==24.6.* 
pylibraft-cu12==24.6.*
raft-dask-cu12==24.6.* 
cuvs-cu12==24.6.*